refactor/reorganize investigation/exploration into actions folder
