+++ src/Menu/actions/investigation/AddToExplorationsButton.test.tsx
