+++ src/Menu/actions/investigation/AddToExplorationsButton.tsx
@@ -14,9 +14,9 @@ import React from 'react';
 import MimirLogo from 'img/logo.svg';
 import { findObjectOfType } from 'utils';
 
-import pluginJson from '../plugin.json';
-import { VAR_DATASOURCE_EXPR } from '../shared';
-import { isSceneQueryRunner } from '../utils/utils.queries';
+import pluginJson from '../../../plugin.json';
+import { VAR_DATASOURCE_EXPR } from '../../../shared';
+import { isSceneQueryRunner } from '../../../utils/utils.queries';
 
 export const investigationsPluginId = 'grafana-investigations-app';
 export const extensionPointId = `${pluginJson.id}/investigation/v1`;
